fix: Change DistanceTraveled type (double->int)

+++ Interfaces/IMovable.cs
@@ -9,6 +9,6 @@ namespace LR1
     internal interface IMovable
     {
         void Move();
-        double DistanceTraveled { get; }
+        int DistanceTraveled { get; }
     }
 }
